requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
docopt==0.6.2
idna==3.4
pipreqs==0.4.13
pyreqs==0.1.1
requests==2.28.2
sh==2.0.3
urllib3==1.26.15
yarg==0.1.9
bcrypt==4.0.1
cffi==1.15.1
cryptography==40.0.1
greenlet==2.0.2
numpy==1.24.2
pandas==1.5.3
paramiko==3.1.0
psycopg2==2.9.5
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2023.2
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.46
sshtunnel==0.4.0
utils==1.0.1bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
docopt==0.6.2
geographiclib==2.0
geopy==2.3.0
googletrans==4.0.0rc1
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
numpy==1.24.2
pandas==1.5.3
paramiko==3.1.0
pipreqs==0.4.13
psycopg2==2.9.5
pycparser==2.21
PyNaCl==1.5.0
pyreqs==0.1.1
python-dateutil==2.8.2
pytz==2023.2
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
sh==2.0.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
sshtunnel==0.4.0
urllib3==1.26.15
utils==1.0.1
yarg==0.1.9
